Added new grip type, Custom Grip, it does not perform any movement logic and sends an interface event called TickGrip every tick with the motion controllers delta since last frame.
TickGrip can be enabled for ALL grip types as a boolean in the motion controller component.

Fixed a couple minor bugs

Change main tick function into a switch statement

Moved interface checks before all logic in tick and set booleans to prevent having to check
for the interface multiple times.

Deleted a ton of commented code in the motion controller that isn't going to be used
in the future.

--HG--
branch : Master