Made base characters and base movement components hold more of the logic for the VR characters
This change allows blueprint characters to be abstract between VRChar and SimpleVRChar and switch between the two (so that I don't have to duplicate nodes and code
between the two base classes when working on my template).

It also causes less code duplication and clarifies class roles more.

This change moves the VRForwardVector / Right / Location functions into the character itself for the VRCharacter, like the simple character already had it.

--HG--
branch : 4.15 Alpha