Added preprocessor definitions to the plugin modules so c++ can control different code paths depending on if they are enabled or not.
Added bIgnoreSimulatingComponentsInFloorCheck variable
to the base movement component. When true simulating components
are never considered for floor checks (to avoid instability / physics issues).

This adds parity for this feature to the simpleCharacter, it was
vrCharacter specific before. It also allows for it to be toggled off if wanted.

Added flex note for users of it to build.cs (uncomment to enable FLEX)

Removed a couple of redundent dependancy modules

--HG--
branch : Master