Final refactor push before 4.16
Added GoogleVR to HMD Types in GetHMDType to keep sync with Epics

Finally was able to convert OpenVRExpansionFunctionLibrary to a blueprint library only now
that it is in its own module. Uses a static OpenVRGenericInterface just like SteamVRHMD.

Added more notes for incoming 4.16 upgrade.

Added SteamVRKeyboardComponent that allows for displaying, interacting with, and getting
data back from the SteamVR Keyboard while in game. Positions itself with the component
wherever it is and polls the events from the keyboard through an overlay it creates to host it.

Component has several events that can be Binded to in Blueprint and the OpenVRKeyboard and CloseVRKeyboard
functions.

--HG--
branch : Master