Removed custom NetSerialize functions on all structs except for the ReplicatedPos struct (as that changes every send anyway and needs to full rep).
I was unaware that setting a custom NetSerialize prevented the delta serialization of
structs from working. After network profiling however this appears to be the case, I had assumed
that NetSerialize covered initial and full serialization only.

Re-worked how the OnRep of the gripped array's detects significant changes to
a grip that require re-initilization. Now that delta is working again the
previous array value that gets passed in is of the previous delta, not the previous array.

Because of this I had to move to caching important values and cross checking them with the new
replicated values instead.

Set all grippable components back to the default of bReplicates = false. On these
the VRGripInterfaceSettings will not replicate unless you set the component to replicate. This
makes more sense, and the grippable actors themselves which default to bReplicates = true
auto replicate their VRGripInterfaceSettings.

Set position and rotation to not replicate on the VRCamera and GripMotionControllers unless
either has changed. This doesn't do much when in VR as they pretty much always change. However
when testing with FPS pawns this is a bandwidth savings.

--HG--
branch : 4.16