Added GetVRDevice_Matrix34Property function to the OpenVR Library, returns them as FTransforms.
Changed GetClosestSlotInRange to check for snap range off of relative location for the snap distance
instead of worldspace. This way snapping ranges should be object relative regardless of scale.

GetClosestSlotInRange is now returning a worldtransform with an identity scale instead of the default
which contains the actor scale. This way when made relative to the object the transform will
be correct regardless of scaling.

Physics simulating objects now are manually scaled if Scaling_Retain secondary grip type
is defined in their VRGripInterface, this brings them to par with non simulating objects.
Since simulating objects are not directly moved, but rather constrained to a kinematic object
it was nessecary to manually rescale them seperately if the grip called for it. Current implementation
may be streamlined in the future.

**TEMPLATE BUT RELEVANT**
Template Gun logic for snap point location check modified to support the fact that if
scaled now it will have a modified relative location. (Need to consider whether to support
direct grip storage of slotted grip or to continue leaving it up to users). Original logic
will continue to function for 1:1 scaled objects, but up/downscaled objects would need to
adjust.

If I add direct support for "was slot" I would have to add the name of the slot as well for identification.

I don't really want to bloat the grip structures or interface structures any more than they already are.

--HG--
branch : Master