Added prelim support for replicating capsule half heights in multiplayer. This is controllable by a new boolean VRReplicateCapsuleHeight on the character movement component.
It causes minor issues with climbing step up and edge cases currently.

Also now zeroing ALL Z values for HMD relative movement for the movement system.
Before I was only doing it for the simple character.

This free's up the Z axis value to be used for the capsule half height replication
so that I don't have to pass any more variables on.

--HG--
branch : Master