Quick fix for bUseWalkingCollisionOverride and Stepup
Another quick fix for an issue that cropped up in 4.16 with bUseWalkingCollisionOverride

I'll need to do truly custom logic for the mode soon, its not acceptable for it to break
when Epic makes changes. Will likely go in soon and start branching movement logic
based on the boolean setting as well as correctly handling the HMD input instead of
just adding to the input vector.

--HG--
branch : Master