Added GripControllerIsTracked blueprint function to act as a replacement for IsTracked() in blueprint.
This is less than ideal but I have literally no access to any of the variables or functions
that the base class uses for it.

I don't want to go back to the controllers being their own class due to the new tracked device
control library that Epic added.

--HG--
branch : Master