Added calling controller capability to the GetClosestSocketInRange interface functions. Can be used in overridden versions to check the hand / other info about what/who is checking for a socket in range.
Defaults to a nullptr, but a motion controller can be passed in and used on the implementation side.

Template will have it filled in.


Made bHadRelativeMovement publically readable in blueprints
While misnamed, is true if we collided with a wall/obstacle due to the HMDs movement in this frame (not movement components)


Removed MotionControllerLocDelta from TickGrip
Not only can you manually calc it (and probably should) with custom grips.
But being in worldspace the offset could be wrong due to character movement.

Better to let items calc it themselves and store the variables for more advanced use instead.
I don't think it had much use in its current state.


Added DoNotMoveIgnoreEntireHeldActor to advanced physics settings.
This allows for gripping drawers and doors without removing capsule collision to the entire actor
for the person gripping.
Set DoNotMoveIgnoreEntireHeldActor to true for all doors and drawers in example content.

Added somewhat working dial component - needs full testing

--HG--
branch : Master