Removed DoNotMoveIgnoreEntireHeldActor
I had a better solution come up. Now I am only calling move ignore actor if the held
object is an entire actor. If the grip is on a component only it does not perform
it for the entire actor, just for the held component.

If for whatever reason someone needs an entire actor to be uncollidable with a component
grip they can do it themselves.

--HG--
branch : Master