Added the following two variables to the DialComponent
ClockwiseMaximumDialAngle
CClockwiseMaximumDialAngle

They control how far in each direction the dial can rotate from 0, default is 180 for both (360 total).

Made dial correctly output as 0-360 clamped angle instead of allowing for negative values
and multiple revolutions.

Setting one of the MaximumDialAngles to 0.0 will cause it to not be able to rotate in that direction.

--HG--
branch : Master