Fixed a bug with HasGripAuthority where the server was being assigned as grip authority for the Client_Auth grip on its end.
Added a 1 frame delay flag after teleport where grips won't check the constraint for
if it is outside of the auto drop distance. This is to allow for the constraints localpose to
held object to be recalculated. Otherwise it could rarely drop the held object on teleport with
small auto drop lengths.

--HG--
branch : Master