Added replication smoothing to the camera like the controllers have
Made the first replication directly set the location instead of beginning a lerp, all other replications
after that are lerped too. This prevents the initial 'tick' of lerp going from 0,0,0.

Set the VivePawnCharacter to default smooth the camera replication and to a 60htz rate instead
of the 100htz default.

--HG--
branch : Master