Added a bunch of functions/nodes to the OpenVR library to control overlay skyboxs and loading systems. They are detailed in the wiki entry below.
https://bitbucket.org/mordentral/vrexpansionplugin/wiki/VRExpansionFunctionLibrary

Corrected an issue that the Epic Motion Controllers have that was evident in my overrides of them, now removing the late update scene view if the motion
controller is deactivated. Otherwise late updates would continue to be used and visually it would appear as if the controller was still active.

--HG--
branch : Master