Fixed some missing includes
Working multidirectional lever

Has some changes to how it behaves in that mode though

For one it only (currently) works with LerpToZero or Stay

Also the angle is always positive and you should use the new variable CurrentLeverForwardVector
to handle its facing direction. (IE: Directon Vector = Direction moved in, Angle = how far held
down in that direction).

CurrentLeverForwardVector is a new variable that has been added, even X/Y modes implement it so it can
be used with them as well.

--HG--
branch : Master