Fixed min distance clamping with the slider (oops).
Also added CurrentSliderProgress variable that is calculated during movement.

Ranges from 0.0 - 1.0 with 1.0 being 100% and 0 being 0%, also works for along the spline
component it is following.

Added SetSplineComponentToFollow function to either add or clear the follow spline, this will also
zero out the initial transform and re-calculate the CurrentSliderProgress.

--HG--
branch : Master