Fixes an issue with the new late update method in 4.18 that I early updated to.
Since it is queuing the late update now to avoid thread racing in game and render,
it applies the final late update the one frame later in render thread
after the object is no longer late updated so if you don't move or update the
component (or its parent) after releasing it from the late update then it never
loses the rendering offset.

I am just calling UpdateComponentToWorld now on release to correct for it.

--HG--
branch : Master