Moved the HMD and controller RPCS into the character itself using a function pointer.
This way if someone uses the controllers without my characters they will still function.
However this saves about 8 bits of data PER object PER RPC. Significant enough to do it.

Epic also does this in 4.19 for the character movement RPCs now.


Fixed some camera stuff and how epic treats the camera, before Vives were snapping back
to point 0,0,0 when they lost tracking, which is bad for this kind of thing.

they recently fixed the controllers, but not the HMDs for this.

--HG--
branch : Master