Added PerformMoveAction_SnapTurn and PerformMoveAction_Teleport as well as the start of a generic MoveAction system.
This provides a way to have client side instant movement actions that tie in
cleanly with the saved movement system and rollbacks but are still server authed.

(Nothing of this sort is normally supported in engine, and for VR it is really
needed, if only for the teleporting and snap turns).

Will be adding in support for custom ones eventually as well as verification events
to bind to to check for safety.

Rep by Rep cost is -1 bit (new control bool), however when one is used in that movement
the cost goes up to +2 bits over original (+3 bits over current) + the cost of
repping either rot/loc or both for the movement action (customs rep both).

This system is NOT flag based, setting a new action for the frame erases any previous
action, you can only use one per tick.

--HG--
branch : Master