The great VRCharacter refactoring
Added OnBeginWallPushback and OnEndWallPushback events to the characters so people can
darken vision and set up things when pushback is happening.

Added vr.ForceNoStereoWithVRWidgets so that users can globally force stereo widgets
to use only normal wiget rendering at will.

Major overhaul of how the VRcharacter handles movement, cleans up many many interactions
and fixes the few broken features from the old system (bCanWalkOffLedges works now).
Also wall sliding is far cleaner and nicer now.

Basically instead of the old hacky method of adding a bit of movement input
in the normal direction of HMD movement (that I hated and proved more stable than
I had expected). Now during movement the movement mode rewinds the last delta of the
HMD and plays it back as part of the current movement mode. This allows it to act
as a true first party in the movement logic and keeps all of the normal character
interactions largly as they would be in a 2D pawn.

This is a beta branch, while it has been tested in multiplayer and singleplayer
I need to run it through its paces as much as possible since it is replacing the
core system of the primary character.

That being said, so far it appears to be more accurate, smoother, and far more reliable
in multiplayer.

I will likely start stripping out root motion code anywhere I can soon as it is
useless in VR characters and will only slow down processes.

Also added temp class NetQuantize1000, don't expect this to be in the final push.
I need it for precision checks currently.

--HG--
branch : VRCharacter Refactor 4.18