Climbing no longer hitches on using it and dropping it right away while standing on the floor
Added -2.15f Z offset to VRRootCapsule by default to sink it into the ground
a little bit and counteract the walking hover.

Adding the walking hover to falling mode (checking for floor during falling).
This removes most of the hitching when walking off of objects and entering into
falling directly out of climbing. (this appears to have been an Epic Games Bug
as it appears in all of their games using a character and in thier templates).

Zeroing out the velocity of the character if an auto step up is performed.
This keeps custom step up movement modes from receiving the last known movement
velocity.

--HG--
branch : Master