Fixed teleport move grip - was missing a UPARAM(ref) signature
Default rotation replication to shorts for the VRCharacters

This removes stuttering from lack of precision on rotation operations.
For VR we pretty much always have 0 roll and 0 pitch anyway so this only costs 1 byte per rep.

Which is an overhead, but it pales in comparison to the other replication costs of VR.

You can manually set it back to 1 byte rep if you don't use turning much (or smooth turning)
or don't notice issues with it. However it is defaulted now because the benifits and early out of bug
testing outweighs the downsides.

--HG--
branch : Master