Adding log message warning of interactible setting deprecation Think 4.20 is the patch that I finally remove this capability and force interactibles and custom grips instead.
Added events that are thrown on the motion controllers for when an object is
gripped and dropped. This is useful for feedback like hand attachment and
hiding or specific operations based on what was gripped.

Renamed DenyGripping interface event to "IsDenyingGrips" for blueprints.
Having the node show up as a callable node was not considered when it was originally added.
The function name remains the same so C++ should require no changes.

--HG--
branch : Master