Changed SecondaryRelativeLocation into SecondaryRelativeTransform, this is to allow for more complex secondary grips in the future and things like forehand tracking during a grip.
Converted physics grips to use the same pathway regardless of type (no longer a seperate
logic branch for manipulation grips, now they just use different defualts).

Removed the bDoNotSetCOM option from the advanced physics settings.

Replaced it with an enum with a bunch of seperate options that also now apply to the
manipulation grip. These extra options are in the AdvancedPhysics section of the grip
settings.

/* Use the default setting for the specified grip type */
COM_Default = 0,
/* Don't grip at center of mass (generally unstable as it grips at actor zero)*/
COM_AtPivot = 1,
/* Set center of mass to grip location and grip there (default for interactible with physics) */
COM_SetAndGripAt = 2,
/* Grip at center of mass but do not set it */
COM_GripAt = 3,
/* Just grip at the controller location, but don't set COM (default for manipulation grips)*/
COM_GripAtControllerLoc = 4

These allow for more complete physics control of held objects, as well as makes it easier to
have multiple physics grips on a single object (future patch is going to give IsHeld an array
of controllers instead of returning a single one).

This has the side effect of slightly lowering memory (1 transform) and slightly lowering
CPU costs of physics grips (less transform logic and checks) as well as greatly improving the
flexibility of the physics grips.

Added grippable actors to the grippables prefabs.

Lets you change out the root component at will to use skeletal or static mesh components.
This is useful for making a base class and changing out the type of root component further down.
It was always possible to do in BP, but this makes it easier by pre-defining it.

I suggest using the plugins OptionalRep(Static/Skeletal)MeshComponents as the roots
as they implement the optional movement replication and you will be missing the
StaticMeshComponentReplicatesMovement option that StaticMeshActor adds in.

--HG--
branch : Master