Adding in bOffsetByControllerProfile to the motion controller.
If true, it will offset the controllers by the currently loaded profiles
OffsetTransform.

Throws the event OnControllerProfileTransformChanged to the controllers
so that things like procedural meshes can be offset to account for it (proc
meshes load already aligned to the controller type, so they need to be adjusted).


This makes it far easier to add and use controller profiles, besides  the procedural
mesh re-alignment (which most don't use anyway) it will be fairly seamless.

The example template has example profiles now and implementations of all of this.

--HG--
branch : Master