Fixed some string property sizes in the OpenVRExpansionFunctionLibrary
Added OnPlayerStateReplicated bindable event to the VRCharacters, will fire off when the
player state has replicated over (every time it is changed, if you only want it once
then unbind after it is done). Useful for initial player set up on clients.

Corrected a fatal crash issue with how Epic is currently checking motion controllers
for IsLocallyOwned.

Fixed SecondaryGripTypes with the Retain option being broken in some situations when
using client authoritive grips.

Now only setting simulating server side on drop of a non client authority grip to avoid
a replication race condition.

Improved the network overhead of client authoritive grips by replacing some of the
RPC properties with GripIDs now that GripIDs are standard. Also am now using a seperate
drop function for client auth grips instead of re-using the modification one to save
on replication cost.

Moved the SetIsHeld call to before calling OnGrip so that there is a valid IsHeld during
operations in that event.

Added bSetOwnerOnGrip to AdvancedGripSettings, when true (default) gripping an actor
will automatically set its owner to the motion controllers owner (if it is a pawn).
(does not apply to component grips)

On grip release it DOES NOT RESET OWNER, keep this in mind


Added bAllowIgnoringAttachOnOwner to all Grippable *actor* types, when true (default)
the actor will ignore replicated attachment sends when the grippables GripMovementType
is set to Client_Authority or Client_Authority_NoRep, this is to prevent the replication
from interfering with client side control of attachment in these modes.

Overrode OnRep_ReplicateMovement to bypass them not setting the simulated state
if the movement replication had been off for awhile and the state had toggled in the meantime.


Added DropAndSocketGrip/Object functions to the grip controllers. This dedicated function
will drop a grip and attach it to the passed in scene component with the transform / socket
that was given.

Added bool bWasSocketed to the OnGripRelease and OnChildGripRelease events in the GripInterface.
When true the grip was dropped with a socketing function.

Added RequestsSocketing function to the interface, allows querying if the object wants to be socketed.
(On release / general queries), by default has no functionality, override it to return a value.

--HG--
branch : Master