Defaulting welding to true for plugin attachment operations, didn't realize that the default setting is inverse between c++ and BP.
Generally if someone is asking for a socket operation or seated mode, they WANT it to happen.
Don't feel like it should be denied because of simulation (also seated in simulation could be fun).

I'll leave it to the end user to filter by simulation if they choose.

--HG--
branch : 4.19 Locked