Removed rollback on custom movement modes
This is to remove "climbing mode" hanging in place and seated mode issues.

Really only an issue in really high packet loss situations, but wanted to cover it.

Also added a GetReplicatedMovementMode function that returns the current Conjoined movement
mode.

Set OnClimbingSteppedUp to a BlueprintNativeEvent so that c++ users could use it.

--HG--
branch : Master