Exposed the GripMotionControllers DefaultGripScript pointer to blueprints so that people can access it and set variables on it (if it is a blueprint base script that is).
Also Changed the default and gun tools grip scripts so that the pivot point is attained
from the ParentTransform instead of getting the controller location again. This is not
only slightly faster, but it also allows people to override the parent transform with something
else (say a remote component) so that the grip doesn't have to be based on the controller
itself.

(force/remote grips for instance could provide a proxy components transform instead)

--HG--
branch : Master