Took back 1 of the 8 bits that I saved everyone yesterday in order to add a flag for quantization of the rotation on controllers / cameras.
You can now set it from its default Short quantization to a 10 bit per axis
quantization method, effectivly saving 2.5 bytes per rep and with 4x the granularity
of the engine BYTE rotator quantization.

This makes up for the lack of precision on the BYTE method but still allows for some
decent savings.

Also dropped a byte out of the CustomVRInputVector and RequestedVelocityVector.
41,943.04 UU's should be MORE than enough leg room for these values.

Also am now rounding MoveAction_Custom vector values to 2 digit precision. I was
already doing this for all built in movement actions but the custom ones should
also be restricted like this.

--HG--
branch : Master