Added scoped movement to climbing step up call to let us revert a step up.
Added OnPerformClimbingStepUp assignable multicast delegate to the movement component.
If bound to then the movement component will rollback the climbing step up and broadcast
the event instead with the target step up location.

This lets end users override the default teleport stepup with a lerp or something instead.

If not binding to the event then the standard step up logic is performed.

A breif example has been added to the template show casing just implementing a teleport
operation like the default implementation.

--HG--
branch : Master